requirements.txt
Django==3.2.6
djangorestframework==3.12.4
python-decouple==3.4
django-cors-headers==3.7.0
factory-boy==3.2.0
Faker==8.10.3
Pillow==8.3.1
